se agregaron campos can
