+++ src/main/java/io/ReportarEnLinea/client/ApiCallback.java
@@ -1,0 +1,17 @@
+package io.ReportarEnLinea.client;
+
+import java.io.IOException;
+
+import java.util.Map;
+import java.util.List;
+
+public interface ApiCallback<T> {
+    
+    void onFailure(ApiException e, int statusCode, Map<String, List<String>> responseHeaders);
+    
+    void onSuccess(T result, int statusCode, Map<String, List<String>> responseHeaders);
+    
+    void onUploadProgress(long bytesWritten, long contentLength, boolean done);
+    
+    void onDownloadProgress(long bytesRead, long contentLength, boolean done);
+}
