+++ src/main/java/io/ReportarEnLinea/client/ApiClient.java
@@ -1,0 +1,464 @@
+package io.ReportarEnLinea.client;
+
+import org.threeten.bp.LocalDate;
+import org.threeten.bp.OffsetDateTime;
+
+import java.io.File;
+import java.io.IOException;
+import java.io.UnsupportedEncodingException;
+import java.lang.reflect.Type;
+import java.net.URLEncoder;
+import java.security.cert.CertificateException;
+import java.util.ArrayList;
+import java.util.Collection;
+import java.util.Date;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+import java.util.Map.Entry;
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
+
+import javax.net.ssl.HostnameVerifier;
+import javax.net.ssl.SSLContext;
+import javax.net.ssl.SSLSession;
+import javax.net.ssl.SSLSocketFactory;
+import javax.net.ssl.TrustManager;
+import javax.net.ssl.X509TrustManager;
+
+import okhttp3.Call;
+import okhttp3.MediaType;
+import okhttp3.OkHttpClient;
+import okhttp3.OkHttpClient.Builder;
+import okhttp3.Request;
+import okhttp3.RequestBody;
+import okhttp3.Response;
+import okhttp3.internal.http.HttpMethod;
+
+public class ApiClient {
+
+	private String basePath = "";
+	private Map<String, String> defaultHeaderMap = new HashMap<String, String>();
+	private String tempFolderPath = null;
+	private JSON json;
+
+	private OkHttpClient httpClient;
+
+	public ApiClient() {
+		httpClient = new OkHttpClient();
+		json = new JSON();
+		setUserAgent("APIHub-Codegen/1.0.0/java");
+	}
+
+	public String getBasePath() {
+		return basePath;
+	}
+
+	public ApiClient setBasePath(String basePath) {
+		this.basePath = basePath;
+		return this;
+	}
+
+	public OkHttpClient getHttpClient() {
+		return httpClient;
+	}
+
+	public ApiClient setHttpClient(OkHttpClient httpClient) {
+		this.httpClient = httpClient;
+		return this;
+	}
+
+	public ApiClient setUserAgent(String userAgent) {
+		addDefaultHeader("User-Agent", userAgent);
+		return this;
+	}
+
+	public ApiClient addDefaultHeader(String key, String value) {
+		defaultHeaderMap.put(key, value);
+		return this;
+	}
+
+	public String getTempFolderPath() {
+		return tempFolderPath;
+	}
+
+	public ApiClient setTempFolderPath(String tempFolderPath) {
+		this.tempFolderPath = tempFolderPath;
+		return this;
+	}
+
+	@SuppressWarnings("rawtypes")
+	public String parameterToString(Object param) {
+		if (param == null) {
+			return "";
+		} else if (param instanceof Date || param instanceof OffsetDateTime || param instanceof LocalDate) {
+			String jsonStr = json.serialize(param);
+			return jsonStr.substring(1, jsonStr.length() - 1);
+		} else if (param instanceof Collection) {
+			StringBuilder b = new StringBuilder();
+			for (Object o : (Collection) param) {
+				if (b.length() > 0) {
+					b.append(",");
+				}
+				b.append(String.valueOf(o));
+			}
+			return b.toString();
+		} else {
+			return String.valueOf(param);
+		}
+	}
+
+	public List<Pair> parameterToPair(String name, Object value) {
+		List<Pair> params = new ArrayList<Pair>();
+
+		if (name == null || name.isEmpty() || value == null || value instanceof Collection)
+			return params;
+
+		params.add(new Pair(name, parameterToString(value)));
+		return params;
+	}
+
+	@SuppressWarnings("rawtypes")
+	public List<Pair> parameterToPairs(String collectionFormat, String name, Collection value) {
+		List<Pair> params = new ArrayList<Pair>();
+
+		if (name == null || name.isEmpty() || value == null || value.isEmpty()) {
+			return params;
+		}
+
+		if ("multi".equals(collectionFormat)) {
+			for (Object item : value) {
+				params.add(new Pair(name, escapeString(parameterToString(item))));
+			}
+			return params;
+		}
+
+		String delimiter = ",";
+
+		if ("ssv".equals(collectionFormat)) {
+			delimiter = escapeString(" ");
+		} else if ("tsv".equals(collectionFormat)) {
+			delimiter = escapeString("\t");
+		} else if ("pipes".equals(collectionFormat)) {
+			delimiter = escapeString("|");
+		}
+
+		StringBuilder sb = new StringBuilder();
+		for (Object item : value) {
+			sb.append(delimiter);
+			sb.append(escapeString(parameterToString(item)));
+		}
+
+		params.add(new Pair(name, sb.substring(delimiter.length())));
+
+		return params;
+	}
+
+	public String sanitizeFilename(String filename) {
+		return filename.replaceAll(".*[/\\\\]", "");
+	}
+
+	public boolean isJsonMime(String mime) {
+		String jsonMime = "(?i)^(application/json|[^;/ \t]+/[^;/ \t]+[+]json)[ \t]*(;.*)?$";
+		return mime != null && (mime.matches(jsonMime) || mime.equals("*/*"));
+	}
+
+	public String selectHeaderAccept(String[] accepts) {
+		if (accepts.length == 0) {
+			return null;
+		}
+		for (String accept : accepts) {
+			if (isJsonMime(accept)) {
+				return accept;
+			}
+		}
+		return StringUtil.join(accepts, ",");
+	}
+
+	public String selectHeaderContentType(String[] contentTypes) {
+		if (contentTypes.length == 0 || contentTypes[0].equals("*/*")) {
+			return "application/text";
+		}
+		for (String contentType : contentTypes) {
+			if (isJsonMime(contentType)) {
+				return contentType;
+			}
+		}
+		return contentTypes[0];
+	}
+
+	public String escapeString(String str) {
+		try {
+			return URLEncoder.encode(str, "utf8").replaceAll("\\+", "%20");
+		} catch (UnsupportedEncodingException e) {
+			return str;
+		}
+	}
+
+	@SuppressWarnings("unchecked")
+	public <T> T deserialize(Response response, Type returnType) throws ApiException {
+		if (response == null || returnType == null) {
+			return null;
+		}
+
+		if ("byte[]".equals(returnType.toString())) {
+			try {
+				return (T) response.body().bytes();
+			} catch (IOException e) {
+				throw new ApiException(e);
+			}
+		}
+
+		String respBody;
+		try {
+			if (response.body() != null)
+				respBody = response.body().string();
+			else
+				respBody = null;
+		} catch (IOException e) {
+			throw new ApiException(e);
+		}
+
+		if (respBody == null || "".equals(respBody)) {
+			return null;
+		}
+
+		String contentType = response.headers().get("Content-Type");
+		if (contentType == null) {
+			contentType = "application/text";
+		}
+		if (isJsonMime(contentType)) {
+			return json.deserialize(respBody, returnType);
+		} else if (returnType.equals(String.class)) {
+			return (T) respBody;
+		} else {
+			throw new ApiException("Content type \"" + contentType + "\" is not supported for type: " + returnType,
+					response.code(), response.headers().toMultimap(), respBody);
+		}
+	}
+
+	public RequestBody serialize(Object obj, String contentType) throws ApiException {
+		if (obj instanceof byte[]) {
+			return RequestBody.create(MediaType.parse(contentType), (byte[]) obj);
+		} else if (obj instanceof File) {
+			return RequestBody.create(MediaType.parse(contentType), (File) obj);
+		} else if (isJsonMime(contentType)) {
+			String content;
+			if (obj != null) {
+				content = json.serialize(obj);
+			} else {
+				content = null;
+			}
+			return RequestBody.create(MediaType.parse(contentType), content);
+		} else {
+			throw new ApiException("Content type \"" + contentType + "\" is not supported");
+		}
+	}
+
+	public File prepareDownloadFile(Response response) throws IOException {
+		String filename = null;
+		String contentDisposition = response.header("Content-Disposition");
+		if (contentDisposition != null && !"".equals(contentDisposition)) {
+			Pattern pattern = Pattern.compile("filename=['\"]?([^'\"\\s]+)['\"]?");
+			Matcher matcher = pattern.matcher(contentDisposition);
+			if (matcher.find()) {
+				filename = sanitizeFilename(matcher.group(1));
+			}
+		}
+
+		String prefix = null;
+		String suffix = null;
+		if (filename == null) {
+			prefix = "download-";
+			suffix = "";
+		} else {
+			int pos = filename.lastIndexOf(".");
+			if (pos == -1) {
+				prefix = filename + "-";
+			} else {
+				prefix = filename.substring(0, pos) + "-";
+				suffix = filename.substring(pos);
+			}
+			if (prefix.length() < 3)
+				prefix = "download-";
+		}
+
+		if (tempFolderPath == null)
+			return File.createTempFile(prefix, suffix);
+		else
+			return File.createTempFile(prefix, suffix, new File(tempFolderPath));
+	}
+
+	public <T> ApiResponse<T> execute(Call call) throws ApiException {
+		return execute(call, null);
+	}
+
+	public <T> ApiResponse<T> execute(Call call, Type returnType) throws ApiException {
+		try {
+			Response response = call.execute();
+			T data = handleResponse(response, returnType);
+			return new ApiResponse<T>(response.code(), response.headers().toMultimap(), data);
+		} catch (IOException e) {
+			throw new ApiException(e);
+		}
+	}
+
+	public <T> T handleResponse(Response response, Type returnType) throws ApiException {
+		if (response.isSuccessful()) {
+			if (returnType == null || response.code() == 204) {
+				if (response.body() != null) {
+					response.body().close();
+				}
+				return null;
+			} else {
+				return deserialize(response, returnType);
+			}
+		} else {
+			String respBody = null;
+			if (response.body() != null) {
+				try {
+					respBody = response.body().string();
+				} catch (IOException e) {
+					throw new ApiException(response.message(), e, response.code(), response.headers().toMultimap());
+				}
+			}
+			throw new ApiException(response.message(), response.code(), response.headers().toMultimap(), respBody);
+		}
+	}
+
+	public Call buildCall(String path, String method, List<Pair> queryParams, List<Pair> collectionQueryParams,
+			Object body, Map<String, String> headerParams, Map<String, Object> formParams, String[] authNames,
+			ProgressRequestBody.ProgressRequestListener progressRequestListener) throws ApiException {
+		Request request = buildRequest(path, method, queryParams, collectionQueryParams, body, headerParams, formParams,
+				authNames, progressRequestListener);
+
+		return httpClient.newCall(request);
+	}
+
+	public Request buildRequest(String path, String method, List<Pair> queryParams, List<Pair> collectionQueryParams,
+			Object body, Map<String, String> headerParams, Map<String, Object> formParams, String[] authNames,
+			ProgressRequestBody.ProgressRequestListener progressRequestListener) throws ApiException {
+
+		final String url = buildUrl(path, queryParams, collectionQueryParams);
+		final Request.Builder reqBuilder = new Request.Builder().url(url);
+		processHeaderParams(headerParams, reqBuilder);
+
+		String contentType = (String) headerParams.get("Content-Type");
+		// ensuring a default content type
+		if (contentType == null) {
+			contentType = "application/text";
+		}
+
+		RequestBody reqBody;
+		if (!HttpMethod.permitsRequestBody(method)) {
+			reqBody = null;
+		} else if (body == null) {
+			if ("DELETE".equals(method)) {
+				reqBody = null;
+			} else {
+				reqBody = RequestBody.create(MediaType.parse(contentType), "");
+			}
+		} else {
+			reqBody = serialize(body, contentType);
+		}
+
+		Request request = null;
+
+		if (progressRequestListener != null && reqBody != null) {
+			ProgressRequestBody progressRequestBody = new ProgressRequestBody(reqBody, progressRequestListener);
+			request = reqBuilder.method(method, progressRequestBody).build();
+		} else {
+			request = reqBuilder.method(method, reqBody).build();
+		}
+
+		return request;
+	}
+
+	public String buildUrl(String path, List<Pair> queryParams, List<Pair> collectionQueryParams) {
+		final StringBuilder url = new StringBuilder();
+		url.append(basePath).append(path);
+
+		if (queryParams != null && !queryParams.isEmpty()) {
+			String prefix = path.contains("?") ? "&" : "?";
+			for (Pair param : queryParams) {
+				if (param.getValue() != null) {
+					if (prefix != null) {
+						url.append(prefix);
+						prefix = null;
+					} else {
+						url.append("&");
+					}
+					String value = parameterToString(param.getValue());
+					url.append(escapeString(param.getName())).append("=").append(escapeString(value));
+				}
+			}
+		}
+
+		if (collectionQueryParams != null && !collectionQueryParams.isEmpty()) {
+			String prefix = url.toString().contains("?") ? "&" : "?";
+			for (Pair param : collectionQueryParams) {
+				if (param.getValue() != null) {
+					if (prefix != null) {
+						url.append(prefix);
+						prefix = null;
+					} else {
+						url.append("&");
+					}
+					String value = parameterToString(param.getValue());
+					url.append(escapeString(param.getName())).append("=").append(value);
+				}
+			}
+		}
+
+		return url.toString();
+	}
+
+	public void processHeaderParams(Map<String, String> headerParams, Request.Builder reqBuilder) {
+		for (Entry<String, String> param : headerParams.entrySet()) {
+			reqBuilder.header(param.getKey(), parameterToString(param.getValue()));
+		}
+		for (Entry<String, String> header : defaultHeaderMap.entrySet()) {
+			if (!headerParams.containsKey(header.getKey())) {
+				reqBuilder.header(header.getKey(), parameterToString(header.getValue()));
+			}
+		}
+	}
+
+	@SuppressWarnings("deprecation")
+	public static OkHttpClient getClientNoSSLVerification() {
+		try {
+			final TrustManager[] trustAllCerts = new TrustManager[] { new X509TrustManager() {
+				@Override
+				public void checkClientTrusted(java.security.cert.X509Certificate[] chain, String authType)
+						throws CertificateException {
+				}
+
+				@Override
+				public void checkServerTrusted(java.security.cert.X509Certificate[] chain, String authType)
+						throws CertificateException {
+				}
+
+				@Override
+				public java.security.cert.X509Certificate[] getAcceptedIssuers() {
+					return new java.security.cert.X509Certificate[] {};
+				}
+			} };
+
+			final SSLContext sslContext = SSLContext.getInstance("SSL");
+			sslContext.init(null, trustAllCerts, new java.security.SecureRandom());
+			final SSLSocketFactory sslSocketFactory = sslContext.getSocketFactory();
+			Builder okHttpClient = new OkHttpClient().newBuilder().sslSocketFactory(sslSocketFactory)
+					.hostnameVerifier(new HostnameVerifier() {
+						@Override
+						public boolean verify(String hostname, SSLSession session) {
+							return true;
+						}
+					});
+
+			return okHttpClient.build();
+		} catch (Exception e) {
+			throw new RuntimeException(e);
+		}
+	}
+
+}
