+++ src/main/java/io/ReportarEnLinea/client/ApiException.java
@@ -1,0 +1,54 @@
+package io.ReportarEnLinea.client;
+
+import java.util.Map;
+import java.util.List;
+
+
+public class ApiException extends Exception {
+    private int code = 0;
+    private Map<String, List<String>> responseHeaders = null;
+    private String responseBody = null;
+    public ApiException() {}
+    public ApiException(Throwable throwable) {
+        super(throwable);
+    }
+    public ApiException(String message) {
+        super(message);
+    }
+    public ApiException(String message, Throwable throwable, int code, Map<String, List<String>> responseHeaders, String responseBody) {
+        super(message, throwable);
+        this.code = code;
+        this.responseHeaders = responseHeaders;
+        this.responseBody = responseBody;
+    }
+    public ApiException(String message, int code, Map<String, List<String>> responseHeaders, String responseBody) {
+        this(message, (Throwable) null, code, responseHeaders, responseBody);
+    }
+    public ApiException(String message, Throwable throwable, int code, Map<String, List<String>> responseHeaders) {
+        this(message, throwable, code, responseHeaders, null);
+    }
+    public ApiException(int code, Map<String, List<String>> responseHeaders, String responseBody) {
+        this((String) null, (Throwable) null, code, responseHeaders, responseBody);
+    }
+    public ApiException(int code, String message) {
+        super(message);
+        this.code = code;
+    }
+    public ApiException(int code, String message, Map<String, List<String>> responseHeaders, String responseBody) {
+        this(code, message);
+        this.responseHeaders = responseHeaders;
+        this.responseBody = responseBody;
+    }
+    
+    public int getCode() {
+        return code;
+    }
+    
+    public Map<String, List<String>> getResponseHeaders() {
+        return responseHeaders;
+    }
+    
+    public String getResponseBody() {
+        return responseBody;
+    }
+}
