+++ src/main/java/io/ReportarEnLinea/client/Configuration.java
@@ -1,0 +1,14 @@
+package io.ReportarEnLinea.client;
+
+
+public class Configuration {
+    private static ApiClient defaultApiClient = new ApiClient();
+    
+    public static ApiClient getDefaultApiClient() {
+        return defaultApiClient;
+    }
+    
+    public static void setDefaultApiClient(ApiClient apiClient) {
+        defaultApiClient = apiClient;
+    }
+}
