+++ src/main/java/io/ReportarEnLinea/client/Pair.java
@@ -1,0 +1,30 @@
+package io.ReportarEnLinea.client;
+
+
+public class Pair {
+    private String name = "";
+    private String value = "";
+    public Pair (String name, String value) {
+        setName(name);
+        setValue(value);
+    }
+    private void setName(String name) {
+        if (!isValidString(name)) return;
+        this.name = name;
+    }
+    private void setValue(String value) {
+        if (!isValidString(value)) return;
+        this.value = value;
+    }
+    public String getName() {
+        return this.name;
+    }
+    public String getValue() {
+        return this.value;
+    }
+    private boolean isValidString(String arg) {
+        if (arg == null) return false;
+        if (arg.trim().isEmpty()) return false;
+        return true;
+    }
+}
