+++ src/main/java/io/ReportarEnLinea/client/StringUtil.java
@@ -1,0 +1,24 @@
+package io.ReportarEnLinea.client;
+
+
+public class StringUtil {
+  
+  public static boolean containsIgnoreCase(String[] array, String value) {
+    for (String str : array) {
+      if (value == null && str == null) return true;
+      if (value != null && value.equalsIgnoreCase(str)) return true;
+    }
+    return false;
+  }
+  
+  public static String join(String[] array, String separator) {
+    int len = array.length;
+    if (len == 0) return "";
+    StringBuilder out = new StringBuilder();
+    out.append(array[0]);
+    for (int i = 1; i < len; i++) {
+      out.append(separator).append(array[i]);
+    }
+    return out.toString();
+  }
+}
