--- src/main/java/io/apihub/client/Configuration.java
@@ -1,13 +1,0 @@
-package io.apihub.client;
-
-public class Configuration {
-	private static ApiClient defaultApiClient = new ApiClient();
-
-	public static ApiClient getDefaultApiClient() {
-		return defaultApiClient;
-	}
-
-	public static void setDefaultApiClient(ApiClient apiClient) {
-		defaultApiClient = apiClient;
-	}
-}
