--- src/main/java/io/apihub/client/GzipRequestInterceptor.java
@@ -1,69 +1,0 @@
-package io.apihub.client;
-
-import okio.Buffer;
-import okio.BufferedSink;
-import okio.GzipSink;
-import okio.Okio;
-
-import java.io.IOException;
-
-import okhttp3.Interceptor;
-import okhttp3.MediaType;
-import okhttp3.Request;
-import okhttp3.RequestBody;
-import okhttp3.Response;
-
-class GzipRequestInterceptor implements Interceptor {
-	@Override
-	public Response intercept(Chain chain) throws IOException {
-		Request originalRequest = chain.request();
-		if (originalRequest.body() == null || originalRequest.header("Content-Encoding") != null) {
-			return chain.proceed(originalRequest);
-		}
-		Request compressedRequest = originalRequest.newBuilder().header("Content-Encoding", "gzip")
-				.method(originalRequest.method(), forceContentLength(gzip(originalRequest.body()))).build();
-		return chain.proceed(compressedRequest);
-	}
-
-	private RequestBody forceContentLength(final RequestBody requestBody) throws IOException {
-		final Buffer buffer = new Buffer();
-		requestBody.writeTo(buffer);
-		return new RequestBody() {
-			@Override
-			public MediaType contentType() {
-				return requestBody.contentType();
-			}
-
-			@Override
-			public long contentLength() {
-				return buffer.size();
-			}
-
-			@Override
-			public void writeTo(BufferedSink sink) throws IOException {
-				sink.write(buffer.snapshot());
-			}
-		};
-	}
-
-	private RequestBody gzip(final RequestBody body) {
-		return new RequestBody() {
-			@Override
-			public MediaType contentType() {
-				return body.contentType();
-			}
-
-			@Override
-			public long contentLength() {
-				return -1;
-			}
-
-			@Override
-			public void writeTo(BufferedSink sink) throws IOException {
-				BufferedSink gzipSink = Okio.buffer(new GzipSink(sink));
-				body.writeTo(gzipSink);
-				gzipSink.close();
-			}
-		};
-	}
-}
