--- src/main/java/io/apihub/client/JSON.java
@@ -1,326 +1,0 @@
-package io.apihub.client;
-
-import com.google.gson.Gson;
-import com.google.gson.GsonBuilder;
-import com.google.gson.JsonParseException;
-import com.google.gson.TypeAdapter;
-import com.google.gson.internal.bind.util.ISO8601Utils;
-import com.google.gson.stream.JsonReader;
-import com.google.gson.stream.JsonWriter;
-import com.google.gson.JsonElement;
-import io.gsonfire.GsonFireBuilder;
-import org.threeten.bp.LocalDate;
-import org.threeten.bp.OffsetDateTime;
-import org.threeten.bp.format.DateTimeFormatter;
-
-import okio.ByteString;
-
-import java.io.IOException;
-import java.io.StringReader;
-import java.lang.reflect.Type;
-import java.text.DateFormat;
-import java.text.ParseException;
-import java.text.ParsePosition;
-import java.util.Date;
-import java.util.Map;
-
-public class JSON {
-	private Gson gson;
-	private boolean isLenientOnJson = false;
-	private DateTypeAdapter dateTypeAdapter = new DateTypeAdapter();
-	private SqlDateTypeAdapter sqlDateTypeAdapter = new SqlDateTypeAdapter();
-	private OffsetDateTimeTypeAdapter offsetDateTimeTypeAdapter = new OffsetDateTimeTypeAdapter();
-	private LocalDateTypeAdapter localDateTypeAdapter = new LocalDateTypeAdapter();
-	private ByteArrayAdapter byteArrayAdapter = new ByteArrayAdapter();
-
-	public static GsonBuilder createGson() {
-		GsonFireBuilder fireBuilder = new GsonFireBuilder();
-		GsonBuilder builder = fireBuilder.createGsonBuilder();
-		return builder;
-	}
-
-	@SuppressWarnings("unused")
-	private static String getDiscriminatorValue(JsonElement readElement, String discriminatorField) {
-		JsonElement element = readElement.getAsJsonObject().get(discriminatorField);
-		if (null == element) {
-			throw new IllegalArgumentException("missing discriminator field: <" + discriminatorField + ">");
-		}
-		return element.getAsString();
-	}
-
-	@SuppressWarnings({ "unused", "rawtypes" })
-	private static Class getClassByDiscriminator(Map classByDiscriminatorValue, String discriminatorValue) {
-		Class clazz = (Class) classByDiscriminatorValue.get(discriminatorValue.toUpperCase());
-		if (null == clazz) {
-			throw new IllegalArgumentException("cannot determine model class of name: <" + discriminatorValue + ">");
-		}
-		return clazz;
-	}
-
-	public JSON() {
-		gson = createGson().registerTypeAdapter(Date.class, dateTypeAdapter)
-				.registerTypeAdapter(java.sql.Date.class, sqlDateTypeAdapter)
-				.registerTypeAdapter(OffsetDateTime.class, offsetDateTimeTypeAdapter)
-				.registerTypeAdapter(LocalDate.class, localDateTypeAdapter)
-				.registerTypeAdapter(byte[].class, byteArrayAdapter).create();
-	}
-
-	public Gson getGson() {
-		return gson;
-	}
-
-	public JSON setGson(Gson gson) {
-		this.gson = gson;
-		return this;
-	}
-
-	public JSON setLenientOnJson(boolean lenientOnJson) {
-		isLenientOnJson = lenientOnJson;
-		return this;
-	}
-
-	public String serialize(Object obj) {
-		return gson.toJson(obj);
-	}
-
-	@SuppressWarnings("unchecked")
-	public <T> T deserialize(String body, Type returnType) {
-		try {
-			if (isLenientOnJson) {
-				JsonReader jsonReader = new JsonReader(new StringReader(body));
-				jsonReader.setLenient(true);
-				return gson.fromJson(jsonReader, returnType);
-			} else {
-				return gson.fromJson(body, returnType);
-			}
-		} catch (JsonParseException e) {
-			if (returnType.equals(String.class))
-				return (T) body;
-			else
-				throw (e);
-		}
-	}
-
-	public class ByteArrayAdapter extends TypeAdapter<byte[]> {
-		@Override
-		public void write(JsonWriter out, byte[] value) throws IOException {
-			if (value == null) {
-				out.nullValue();
-			} else {
-				out.value(ByteString.of(value).base64());
-			}
-		}
-
-		@Override
-		public byte[] read(JsonReader in) throws IOException {
-			switch (in.peek()) {
-			case NULL:
-				in.nextNull();
-				return null;
-			default:
-				String bytesAsBase64 = in.nextString();
-				ByteString byteString = ByteString.decodeBase64(bytesAsBase64);
-				return byteString.toByteArray();
-			}
-		}
-	}
-
-	public static class OffsetDateTimeTypeAdapter extends TypeAdapter<OffsetDateTime> {
-		private DateTimeFormatter formatter;
-
-		public OffsetDateTimeTypeAdapter() {
-			this(DateTimeFormatter.ISO_OFFSET_DATE_TIME);
-		}
-
-		public OffsetDateTimeTypeAdapter(DateTimeFormatter formatter) {
-			this.formatter = formatter;
-		}
-
-		public void setFormat(DateTimeFormatter dateFormat) {
-			this.formatter = dateFormat;
-		}
-
-		@Override
-		public void write(JsonWriter out, OffsetDateTime date) throws IOException {
-			if (date == null) {
-				out.nullValue();
-			} else {
-				out.value(formatter.format(date));
-			}
-		}
-
-		@Override
-		public OffsetDateTime read(JsonReader in) throws IOException {
-			switch (in.peek()) {
-			case NULL:
-				in.nextNull();
-				return null;
-			default:
-				String date = in.nextString();
-				if (date.endsWith("+0000")) {
-					date = date.substring(0, date.length() - 5) + "Z";
-				}
-				return OffsetDateTime.parse(date, formatter);
-			}
-		}
-	}
-
-	public class LocalDateTypeAdapter extends TypeAdapter<LocalDate> {
-		private DateTimeFormatter formatter;
-
-		public LocalDateTypeAdapter() {
-			this(DateTimeFormatter.ISO_LOCAL_DATE);
-		}
-
-		public LocalDateTypeAdapter(DateTimeFormatter formatter) {
-			this.formatter = formatter;
-		}
-
-		public void setFormat(DateTimeFormatter dateFormat) {
-			this.formatter = dateFormat;
-		}
-
-		@Override
-		public void write(JsonWriter out, LocalDate date) throws IOException {
-			if (date == null) {
-				out.nullValue();
-			} else {
-				out.value(formatter.format(date));
-			}
-		}
-
-		@Override
-		public LocalDate read(JsonReader in) throws IOException {
-			switch (in.peek()) {
-			case NULL:
-				in.nextNull();
-				return null;
-			default:
-				String date = in.nextString();
-				return LocalDate.parse(date, formatter);
-			}
-		}
-	}
-
-	public JSON setOffsetDateTimeFormat(DateTimeFormatter dateFormat) {
-		offsetDateTimeTypeAdapter.setFormat(dateFormat);
-		return this;
-	}
-
-	public JSON setLocalDateFormat(DateTimeFormatter dateFormat) {
-		localDateTypeAdapter.setFormat(dateFormat);
-		return this;
-	}
-
-	public static class SqlDateTypeAdapter extends TypeAdapter<java.sql.Date> {
-		private DateFormat dateFormat;
-
-		public SqlDateTypeAdapter() {
-		}
-
-		public SqlDateTypeAdapter(DateFormat dateFormat) {
-			this.dateFormat = dateFormat;
-		}
-
-		public void setFormat(DateFormat dateFormat) {
-			this.dateFormat = dateFormat;
-		}
-
-		@Override
-		public void write(JsonWriter out, java.sql.Date date) throws IOException {
-			if (date == null) {
-				out.nullValue();
-			} else {
-				String value;
-				if (dateFormat != null) {
-					value = dateFormat.format(date);
-				} else {
-					value = date.toString();
-				}
-				out.value(value);
-			}
-		}
-
-		@Override
-		public java.sql.Date read(JsonReader in) throws IOException {
-			switch (in.peek()) {
-			case NULL:
-				in.nextNull();
-				return null;
-			default:
-				String date = in.nextString();
-				try {
-					if (dateFormat != null) {
-						return new java.sql.Date(dateFormat.parse(date).getTime());
-					}
-					return new java.sql.Date(ISO8601Utils.parse(date, new ParsePosition(0)).getTime());
-				} catch (ParseException e) {
-					throw new JsonParseException(e);
-				}
-			}
-		}
-	}
-
-	public static class DateTypeAdapter extends TypeAdapter<Date> {
-		private DateFormat dateFormat;
-
-		public DateTypeAdapter() {
-		}
-
-		public DateTypeAdapter(DateFormat dateFormat) {
-			this.dateFormat = dateFormat;
-		}
-
-		public void setFormat(DateFormat dateFormat) {
-			this.dateFormat = dateFormat;
-		}
-
-		@Override
-		public void write(JsonWriter out, Date date) throws IOException {
-			if (date == null) {
-				out.nullValue();
-			} else {
-				String value;
-				if (dateFormat != null) {
-					value = dateFormat.format(date);
-				} else {
-					value = ISO8601Utils.format(date, true);
-				}
-				out.value(value);
-			}
-		}
-
-		@Override
-		public Date read(JsonReader in) throws IOException {
-			try {
-				switch (in.peek()) {
-				case NULL:
-					in.nextNull();
-					return null;
-				default:
-					String date = in.nextString();
-					try {
-						if (dateFormat != null) {
-							return dateFormat.parse(date);
-						}
-						return ISO8601Utils.parse(date, new ParsePosition(0));
-					} catch (ParseException e) {
-						throw new JsonParseException(e);
-					}
-				}
-			} catch (IllegalArgumentException e) {
-				throw new JsonParseException(e);
-			}
-		}
-	}
-
-	public JSON setDateFormat(DateFormat dateFormat) {
-		dateTypeAdapter.setFormat(dateFormat);
-		return this;
-	}
-
-	public JSON setSqlDateFormat(DateFormat dateFormat) {
-		sqlDateTypeAdapter.setFormat(dateFormat);
-		return this;
-	}
-}
