--- src/main/java/io/apihub/client/ProgressResponseBody.java
@@ -1,59 +1,0 @@
-package io.apihub.client;
-
-import okhttp3.MediaType;
-import okhttp3.ResponseBody;
-
-import java.io.IOException;
-
-import okio.Buffer;
-import okio.BufferedSource;
-import okio.ForwardingSource;
-import okio.Okio;
-import okio.Source;
-
-public class ProgressResponseBody extends ResponseBody {
-	public interface ProgressListener {
-		void update(long bytesRead, long contentLength, boolean done);
-	}
-
-	private final ResponseBody responseBody;
-	private final ProgressListener progressListener;
-	private BufferedSource bufferedSource;
-
-	public ProgressResponseBody(ResponseBody responseBody, ProgressListener progressListener) {
-		this.responseBody = responseBody;
-		this.progressListener = progressListener;
-	}
-
-	@Override
-	public MediaType contentType() {
-		return responseBody.contentType();
-	}
-
-	@Override
-	public long contentLength() {
-		return responseBody.contentLength();
-	}
-
-	@Override
-	public BufferedSource source() {
-		if (bufferedSource == null) {
-			bufferedSource = Okio.buffer(source(responseBody.source()));
-		}
-		return bufferedSource;
-	}
-
-	private Source source(Source source) {
-		return new ForwardingSource(source) {
-			long totalBytesRead = 0L;
-
-			@Override
-			public long read(Buffer sink, long byteCount) throws IOException {
-				long bytesRead = super.read(sink, byteCount);
-				totalBytesRead += bytesRead != -1 ? bytesRead : 0;
-				progressListener.update(totalBytesRead, responseBody.contentLength(), bytesRead == -1);
-				return bytesRead;
-			}
-		};
-	}
-}
